// mtn-zm-core-connector/src/domain/SDKClient/types.ts
export type TSDKOutboundTransferRequest = {
    /** @description Transaction ID from the DFSP backend, used to reconcile transactions between the Switch and DFSP backend systems. */
    homeTransactionId: string;
    from: components['schemas']['transferParty'];
    to: components['schemas']['transferParty'];
    amountType: components['schemas']['AmountType'];
    currency: components['schemas']['Currency'];
    amount: components['schemas']['Amount'];
    transactionType: components['schemas']['transferTransactionType'];
    subScenario?: components['schemas']['TransactionSubScenario'];
    note?: components['schemas']['Note'];
    quoteRequestExtensions?: components['schemas']['extensionListEmptiable'];
    transferRequestExtensions?: components['schemas']['extensionListEmptiable'];
    /** @description Set to true if supplying an FSPID for the payee party and no party resolution is needed. This may be useful is a previous party resolution has been performed. */
    skipPartyLookup?: boolean;
};

export type TSDKOutboundTransferResponse = {
    transferId?: components["schemas"]["CorrelationId"];
    /** @description Transaction ID from the DFSP backend, used to reconcile transactions between the Switch and DFSP backend systems. */
    homeTransactionId: string;
    from: components["schemas"]["transferParty"];
    to: components["schemas"]["transferParty"];
    amountType: components["schemas"]["AmountType"];
    currency: components["schemas"]["Currency"];
    amount: components["schemas"]["Amount"];
    transactionType: components["schemas"]["transferTransactionType"];
    subScenario?: components["schemas"]["TransactionSubScenario"];
    note?: components["schemas"]["Note"];
    currentState?: components["schemas"]["transferStatus"];
    quoteId?: components["schemas"]["CorrelationId"];
    getPartiesResponse?: {
        body: Record<string, never>;
        headers?: Record<string, never>;
    };
    quoteResponse?: {
        body: components["schemas"]["QuotesIDPutResponse"];
        headers?: Record<string, never>;
    };
    /** @description FSPID of the entity that supplied the quote response. This may not be the same as the FSPID of the entity which owns the end user account in the case of a FOREX transfer. i.e. it may be a FOREX gateway. */
    quoteResponseSource?: string;
    conversionRequestId?: components["schemas"]["CorrelationId"];
    fxQuoteResponse?: {
        body: components["schemas"]["FxQuotesPostOutboundResponse"];
        headers?: Record<string, never>;
    };
    /** @description FXPID of the entity that supplied the fxQuotes response. */
    fxQuoteResponseSource?: string;
    fulfil?: {
        body: components["schemas"]["TransfersIDPutResponse"];
        headers?: Record<string, never>;
    };
    lastError?: components["schemas"]["transferError"];
    /** @description Set to true if supplying an FSPID for the payee party and no party resolution is needed. This may be useful is a previous party resolution has been performed. */
    skipPartyLookup?: boolean;
};
// ...
export type TSDKTransferContinuationRequest =
    | SDKSchemeAdapter.V2_0_0.Outbound.Types.transferContinuationAcceptParty
    | SDKSchemeAdapter.V2_0_0.Outbound.Types.transferContinuationAcceptQuote
    | components['schemas']['transferContinuationAcceptConversion']
    | components['schemas']['transferContinuationAcceptQuoteOrConversion'];


export type TUpdateTransferDeps = {
    fineractTransaction: {
        fineractAccountId: number;
        totalAmount: number;
        routingCode: string;
        receiptNumber: string;
        bankNumber: string;
    };
    sdkTransferId: number | string;
};

export type TtransferContinuationResponse = SDKSchemeAdapter.V2_0_0.Outbound.Types.transferResponse;

export type TSDKClientDeps = {
    logger: ILogger;
    httpClient: IHTTPClient;
    schemeAdapterUrl: string;
};

export interface ISDKClient {
    initiateTransfer(transfer: TSDKOutboundTransferRequest): Promise<THttpResponse<TSDKOutboundTransferResponse>>;
    updateTransfer(
        transferAccept: TSDKTransferContinuationRequest,
        id: string,
    ): Promise<THttpResponse<TtransferContinuationResponse>>;
}